refactor: split session routes, add Session.__getitem__, fix docstring, REST cleanup
- Split 1067-line session_routes.py into CRUD (846 lines) + message routes
  (246 lines in session_message_routes.py) for maintainability
- Added Session.__getitem__() for dict-style access (session['field'])
- Fixed Session docstring: .history is a snapshot, not a defensive copy
- Removed POST /session/{sid}/delete (redundant, DELETE /session/{sid} exists)
- Changed POST /sessions/bulk-delete to DELETE /sessions/bulk
- Fixed conftest.py test pollution: pre-import core.models before
  test_agent_loop.py stubs it with MagicMock at collection time

All 272 tests pass (13 new + 259 existing). TestClient end-to-end: 28/28.
