+++ app.py
@@ -441,6 +441,10 @@ from routes.session_routes import setup_session_routes
 session_config = {"REQUEST_TIMEOUT": REQUEST_TIMEOUT, "OPENAI_API_KEY": OPENAI_API_KEY, "SESSIONS_FILE": SESSIONS_FILE}
 app.include_router(setup_session_routes(session_manager, session_config, webhook_manager=webhook_manager))
 
+# Session messages (history, export, compact)
+from routes.session_message_routes import setup_session_message_routes
+app.include_router(setup_session_message_routes(session_manager, session_config, webhook_manager=webhook_manager))
+
 # Admin Danger Zone wipes (Settings → System → Danger Zone)
 from routes.admin_wipe_routes import setup_admin_wipe_routes
 app.include_router(setup_admin_wipe_routes(session_manager))
