+++ tests/conftest.py
@@ -32,3 +32,8 @@ if "src.database" not in sys.modules:
     _db.SessionLocal = MagicMock()
     _db.ModelEndpoint = MagicMock()
     sys.modules["src.database"] = _db
+
+# Pre-import core.models before test_agent_loop.py's module-level stubs
+# run (it replaces sys.modules['core.models'] with a MagicMock during
+# collection, which breaks session import in subsequent tests).
+import core.models  # noqa: E402
